requirements.txt
python-fasthtml==0.2.4
langchain==0.2.12
langchain-ollama==0.1.1
ollama==0.3.1